requirements.txt
websocket-client
json
requests
